carDekho/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
langchain>=0.2.16
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
